Update template.tex (#183)
* Update template.tex

Added option to turn coloring of hyperlinks off. Passing values to colorlinks or urlcolor from yaml does work, but I found that only setting colorlinks=false actually rendered the links (urls as well as refernces to tables and figures) in black color. I have not done this before, so I am not sure if this will do the trick...

* default colorlinks to false