Deprecate the biblio_style option and rename the reference_style option
This commit is intended to eliminate an ambiguity linked to the use of
biblio_style as the name of the tf_article function option. In fact,
biblio_style is traditionally used in the rticles package to indicate the
name of the .bst file that will be used as an argument to the
\bibliographystyle command in the LaTeX template.

The new version of tf_article accepts an option that determines the Taylor
& Francis reference style to be used. Until now, the name biblio_style was
used as the name of this option. In this commit, this name is changed to
reference_style. This not only aligns with the terminology used by Francis
& Taylor (“Reference Style”), but also avoids any confusion on the part of
the user.

The tf_article function has also been modified to include a pre_knit
function, in which the presence of the biblio_style in the YAML header is
checked. If present, a warning message is issued concerning the
depreciation of the biblio_style field.